- added command line options - fixed problem that appeared when using iPhone5 in simulators config array - ui_script used ./config-auomation.js instead of config-automation.example.js when ui_script did not exist - moved core  to bin/ - moved capture.js to lib/ - moved example project to example
