+++ bin/unix_instruments.sh
