+++ example/Hello World International.xcodeproj/project.xcworkspace/contents.xcworkspacedata
