+++ example/Hello World International/Hello World International-Info.plist
