+++ example/Hello World International/NLAppDelegate.m
