+++ example/Hello World International/de.lproj/InfoPlist.strings
