+++ example/Hello World International/de.lproj/Localizable.strings
