+++ example/Hello World International/en.lproj/InfoPlist.strings
