+++ example/Hello World International/en.lproj/Localizable.strings
