+++ example/Hello World International/fr.lproj/InfoPlist.strings
