+++ example/Hello World International/fr.lproj/Localizable.strings
