+++ example/Hello World International/fr.lproj/MainStoryboard.storyboard
