+++ example/Hello World International/ja.lproj/InfoPlist.strings
