+++ example/Hello World International/ja.lproj/Localizable.strings
