+++ lib/capture.js
