+++ lib/lib
@@ -1,0 +1,1 @@
+/Users/merlos/Projects/adianteapps/bitbucket/publish/screenshoter/ui-screen-shooter/bin/../lib
